# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MAIS Wellbest API Documentation",
        "description": "Standard OpenAPI documentation for the MAIS Wellbest application.",
        "contact": {
            "email": "admin@rexonindo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "MAIS Wellbest API Server"
        }
    ],
    "paths": {
...
        "/processes": {
            "get": {
                "tags": [
                    "Master Data - Processes"
                ],
                "summary": "Get list of processes",
                "description": "Returns a paginated list of processes.",
                "operationId": "getProcessesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Process"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Master Data - Processes"
                ],
                "summary": "Create a new process",
                "description": "Stores a new process and returns the created process data.",
                "operationId": "storeProcess",
                "requestBody": {
                    "description": "Process data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProcessRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Process created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Process"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/processes/{id}": {
            "get": {
                "tags": [
                    "Master Data - Processes"
                ],
                "summary": "Get process information",
                "description": "Returns process data by ID.",
                "operationId": "getProcessById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Process ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Process"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Master Data - Processes"
                ],
                "summary": "Update existing process",
                "description": "Updates an existing process and returns the updated process data.",
                "operationId": "updateProcess",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Process ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Process data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProcessRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Process updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Process"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Master Data - Processes"
                ],
                "summary": "Delete existing process",
                "description": "Deletes a process record and returns no content.",
                "operationId": "deleteProcess",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Process ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginationLinks": {
                "title": "Pagination Links",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://localhost/api/v1/departments?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://localhost/api/v1/departments?page=5"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://localhost/api/v1/departments?page=2",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "Pagination Meta",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "path": {
                        "type": "string",
                        "example": "http://localhost/api/v1/departments"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 75
                    }
                },
                "type": "object"
            },
            "Department": {
                "title": "Department",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "departmentCode": {
                        "type": "string",
                        "example": "PROD"
                    },
                    "name": {
                        "type": "string",
                        "example": "Production"
                    },
                    "description": {
                        "type": "string",
                        "example": "Main production department",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Process": {
                "title": "Process",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "processCode": {
                        "type": "string",
                        "example": "PROC-01"
                    },
                    "name": {
                        "type": "string",
                        "example": "Cutting"
                    },
                    "description": {
                        "type": "string",
                        "example": "Process of cutting raw materials",
                        "nullable": true
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    }
                },
                "type": "object"
            },
            "StoreProcessRequest": {
                "title": "Store Process Request",
                "required": [
                    "proc_cd",
                    "proc_nm",
                    "dept_cd"
                ],
                "properties": {
                    "proc_cd": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "PROC-02"
                    },
                    "proc_nm": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Assembling"
                    },
                    "descrp": {
                        "type": "string",
                        "example": "Process of assembling parts",
                        "nullable": true
                    },
                    "dept_cd": {
                        "type": "string",
                        "maxLength": 20,
                        "example": "PROD"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}